feat(git): handle initial pushes for new branches
Previously, pushing a newly created branch would fail if the remote tracking branch didn't exist, as `git.log` would throw an error.

This change introduces a `try-catch` block around the `git.log` call to gracefully handle cases where the remote tracking branch is not yet present. If an error occurs, it's assumed to be an initial push, and `git.push --set-upstream` is used to establish the tracking relationship and push the branch.